#23 - 전체 패키지 구조 변경
계층형 아키텍처 구조 변경
캡슐화, 정보은닉
