+++ src/main/java/site/moasis/monolithicbe/common/exception/FieldErrorDetail.java
@@ -1,0 +1,16 @@
+package site.moasis.monolithicbe.common.exception;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.Setter;
+
+@Setter
+@Getter
+@Builder
+@AllArgsConstructor
+public class FieldErrorDetail {
+    private String field;
+    private String code;
+    private String message;
+}
