+++ src/main/java/site/moasis/monolithicbe/configuration/CorsFilterConfig.java
@@ -1,4 +1,4 @@
-package site.moasis.monolithicbe.application.configuration;
+package site.moasis.monolithicbe.configuration;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
