+++ src/main/java/site/moasis/monolithicbe/configuration/JwtConfig.java
@@ -1,4 +1,4 @@
-package site.moasis.monolithicbe.application.configuration;
+package site.moasis.monolithicbe.configuration;
 
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
