+++ src/main/java/site/moasis/monolithicbe/controller/article/ArticleController.java
@@ -1,0 +1,4 @@
+package site.moasis.monolithicbe.controller.article;
+
+public class ArticleController {
+}
