+++ src/main/java/site/moasis/monolithicbe/domain/comment/CommentDto.java
@@ -1,0 +1,14 @@
+package site.moasis.monolithicbe.domain.comment;
+
+import java.util.*;
+
+public record CommentDto (){
+
+    public record CommentOneDto(
+            Optional<Comment> comment) {
+    }
+
+    public record CommentResponseDto(
+            List<Comment> comments) {
+    }
+}
